Fixed float version of Tools.min

+++ src/main/java/com/fantasticsource/tools/Tools.java
@@ -21,7 +21,7 @@ public class Tools
         return result;
     }
 
-    public static double min(float... values)
+    public static float min(float... values)
     {
         float result = Float.MAX_VALUE;
         for (float value : values)
